Memory handling code cleanup (#1111)
* Many old macros such as `dtmalloc`, `dtrealloc`, `dtfree`, `dterrr` etc. that were used only in few places, were eliminated.
* Added utility functions `dt::malloc` and `dt::realloc` that are somewhat smarter than standard `malloc` and `free`: they will attempt to free up memory if the first attempt at allocating fails with errno=12 (ENOMEM).
* `ThreadsafeWritableBuffer` now uses `dt::shared_mutex` for read/write access control (this replaces the old mechanism which was based on atomically incrementing /decrementing a counter). Now it is no longer possible for this object to become deadlocked if `resize()` call throws an exception.

+++ c/datatable.cc
@@ -43,13 +43,13 @@ DataTable::DataTable(Column** cols)
 
 
 
-DataTable* DataTable::delete_columns(int *cols_to_remove, int n)
+DataTable* DataTable::delete_columns(int *cols_to_remove, int64_t n)
 {
   if (n == 0) return this;
   qsort(cols_to_remove, (size_t)n, sizeof(int), _compare_ints);
   int j = 0;
   int next_col_to_remove = cols_to_remove[0];
-  int k = 0;
+  int64_t k = 0;
   for (int i = 0; i < ncols; ++i) {
     if (i == next_col_to_remove) {
       delete columns[i];
